_back/pyproject.toml
[project]
name = "nuxer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi[standard]>=0.115.9",
    "loguru>=0.7.3",
    "notion-client>=2.3.0",
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-jose[cryptography]>=3.4.0",
    "requests>=2.32.3",
    "slack-bolt>=1.22.0",
    "sqlalchemy>=2.0.38",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.8",
]

[tool.ruff]
line-length = 120
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
]
ignore = ["B008", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
